--- models/bigmodel.dlc
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:c543edb32dce392c86526020e7e683c05af81fa667be43420e345f157947cd5e
-size 53640843
