--- models/bigmodel.fixed.dlc
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:e98d98a5a201f696da66467225b31361e70ff54a904ebf4066cbcc2ff1551fb9
-size 53640843
